create components dir to contain esp32_bsdiff component
Before this change, the CMakeLists.txt file that registers ESP-IDF
component was in the root directory. Other files in this repo assume the
name of this component is `esp32_bsdiff`.

However, if the repo is cloned into a working directory with a name
other than `esp32_bsdiff`, then the component won't be found.

To fix this issue, the component CMakeLists.txt file has been moved
into a new `components/esp32_bsdiff`, so it can be found regardless of
the repo working directory.

Signed-off-by: Nick Miller <nick@golioth.io>