bspatch: remove newsize parameter
This commit removes the newsize parameter and associated sanity checks
in bspatch().

The reason for doing this is that on some systems, the new size may not
be known a priori.

Now, the bspatch() loop will run continuously until an error occurs
or the end of the patch stream input has been reached.

Another change is that the "old read" stream callback functions are now
checking that they're not reading/writing past the end of the buffer.
The overall buffer size (oldsize) is passed into the callback via
struct OldCtx. This is primarily a sanity check, to prevent OOB access.

Signed-off-by: Nick Miller <nick@golioth.io>